requirements.txt
apispec==6.8.2
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
flask-smorest==0.46.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==4.0.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.1.1
SQLAlchemy==2.0.43
urllib3==2.5.0
webargs==8.7.0
Werkzeug==3.1.3
